Add typescript definitions for mergeProps module (#1756)
